update the ints to longs

+++ src/main/java/edu/ksu/canvas/impl/ContentMigrationImpl.java
@@ -52,7 +52,7 @@ public class ContentMigrationImpl extends BaseImpl<ContentMigration, ContentMigr
     }
 
     @Override
-    public Optional<ContentMigration> updateCourseContentMigration(Integer id, CreateCourseContentMigrationOptions options) throws IOException {
+    public Optional<ContentMigration> updateCourseContentMigration(Long id, CreateCourseContentMigrationOptions options) throws IOException {
         LOG.debug("updating course content migration");
         String url = buildCanvasUrl("courses/" + options.getDestinationCourseId() + "/content_migrations/" + id.toString(), Collections.emptyMap());
         Response response = canvasMessenger.putToCanvas(oauthToken, url, options.getOptionsMap());
@@ -61,7 +61,7 @@ public class ContentMigrationImpl extends BaseImpl<ContentMigration, ContentMigr
 
     /******************************** USERS ********************************/
     @Override
-    public  Optional<ContentMigration> getUserContentMigration(String userId, Integer id) throws IOException {
+    public  Optional<ContentMigration> getUserContentMigration(String userId, Long id) throws IOException {
         LOG.debug("listing a content migration for the user");
         String url = buildCanvasUrl("users/" + userId + "/content_migrations/" + id.toString(), Collections.emptyMap());
         Response response = canvasMessenger.getSingleResponseFromCanvas(oauthToken, url);
@@ -84,7 +84,7 @@ public class ContentMigrationImpl extends BaseImpl<ContentMigration, ContentMigr
     }
 
     @Override
-    public Optional<ContentMigration> updateUserContentMigration(Integer id, CreateContentMigrationOptions options) throws IOException {
+    public Optional<ContentMigration> updateUserContentMigration(Long id, CreateContentMigrationOptions options) throws IOException {
         LOG.debug("updating user content migration");
         String url = buildCanvasUrl("users/" + options.getSourceCourseId() + "/content_migrations/" + id.toString(), Collections.emptyMap());
         Response response = canvasMessenger.putToCanvas(oauthToken, url, options.getOptionsMap());
@@ -93,7 +93,7 @@ public class ContentMigrationImpl extends BaseImpl<ContentMigration, ContentMigr
 
     /******************************** GROUPS ********************************/
     @Override
-    public  Optional<ContentMigration> getGroupContentMigration(String groupId, Integer id) throws IOException {
+    public  Optional<ContentMigration> getGroupContentMigration(String groupId, Long id) throws IOException {
         LOG.debug("listing a content migration for the group");
         String url = buildCanvasUrl("groups/" + groupId + "/content_migrations/" + id.toString(), Collections.emptyMap());
         Response response = canvasMessenger.getSingleResponseFromCanvas(oauthToken, url);
@@ -116,7 +116,7 @@ public class ContentMigrationImpl extends BaseImpl<ContentMigration, ContentMigr
     }
 
     @Override
-    public Optional<ContentMigration> updateGroupContentMigration(Integer id, CreateContentMigrationOptions options) throws IOException {
+    public Optional<ContentMigration> updateGroupContentMigration(Long id, CreateContentMigrationOptions options) throws IOException {
         LOG.debug("updating group content migration");
         String url = buildCanvasUrl("groups/" + options.getSourceCourseId() + "/content_migrations/" + id.toString(), Collections.emptyMap());
         Response response = canvasMessenger.putToCanvas(oauthToken, url, options.getOptionsMap());
@@ -125,7 +125,7 @@ public class ContentMigrationImpl extends BaseImpl<ContentMigration, ContentMigr
     
     /******************************** ACCOUNTS ********************************/
     @Override
-    public  Optional<ContentMigration> getAccountContentMigration(String accountId, Integer id) throws IOException {
+    public  Optional<ContentMigration> getAccountContentMigration(String accountId, Long id) throws IOException {
         LOG.debug("listing a content migration for the account");
         String url = buildCanvasUrl("accounts/" + accountId + "/content_migrations/" + id.toString(), Collections.emptyMap());
         Response response = canvasMessenger.getSingleResponseFromCanvas(oauthToken, url);
@@ -148,7 +148,7 @@ public class ContentMigrationImpl extends BaseImpl<ContentMigration, ContentMigr
     }
 
     @Override
-    public Optional<ContentMigration> updateAccountContentMigration(Integer id, CreateContentMigrationOptions options) throws IOException {
+    public Optional<ContentMigration> updateAccountContentMigration(Long id, CreateContentMigrationOptions options) throws IOException {
         LOG.debug("updating account content migration");
         String url = buildCanvasUrl("accounts/" + options.getSourceCourseId() + "/content_migrations/" + id.toString(), Collections.emptyMap());
         Response response = canvasMessenger.putToCanvas(oauthToken, url, options.getOptionsMap());
