+++ src/main/java/edu/ksu/canvas/interfaces/ContentMigrationWriter.java
@@ -11,14 +11,14 @@ import java.util.Optional;
 public interface ContentMigrationWriter extends CanvasWriter<ContentMigration, ContentMigrationWriter>{
 
     Optional<ContentMigration> createCourseContentMigration(CreateCourseContentMigrationOptions options) throws IOException;
-    Optional<ContentMigration> updateCourseContentMigration(Integer id, CreateCourseContentMigrationOptions options) throws IOException;
+    Optional<ContentMigration> updateCourseContentMigration(Long id, CreateCourseContentMigrationOptions options) throws IOException;
 
     Optional<ContentMigration> createUserContentMigration(CreateContentMigrationOptions options) throws IOException;
-    Optional<ContentMigration> updateUserContentMigration(Integer id, CreateContentMigrationOptions options) throws IOException;
+    Optional<ContentMigration> updateUserContentMigration(Long id, CreateContentMigrationOptions options) throws IOException;
 
     Optional<ContentMigration> createGroupContentMigration(CreateContentMigrationOptions options) throws IOException;
-    Optional<ContentMigration> updateGroupContentMigration(Integer id, CreateContentMigrationOptions options) throws IOException;
+    Optional<ContentMigration> updateGroupContentMigration(Long id, CreateContentMigrationOptions options) throws IOException;
 
     Optional<ContentMigration> createAccountContentMigration(CreateContentMigrationOptions options) throws IOException;
-    Optional<ContentMigration> updateAccountContentMigration(Integer id, CreateContentMigrationOptions options) throws IOException;
+    Optional<ContentMigration> updateAccountContentMigration(Long id, CreateContentMigrationOptions options) throws IOException;
 }
